build with VS2017

+++ Cube.hpp
@@ -102,7 +102,7 @@ public:
     IndexedTriangleList<TextureEffect::Vertex> GetIndexedTriangleListTex()
     {
         std::vector<TextureEffect::Vertex> verticesTex;
-        for (int i = 0; i < vertices.size(); i++)
+        for (size_t i = 0; i < vertices.size(); i++)
             verticesTex.emplace_back(vertices[i], textureCoords[i]);
         
         return { verticesTex, triangles };
@@ -110,7 +110,7 @@ public:
     IndexedTriangleList<VertexColorEffect::Vertex> GetIndexedTriangleListVC()
     {
         std::vector<VertexColorEffect::Vertex> verticesVC;
-        for (int i = 0; i < vertices.size(); i++)
+        for (size_t i = 0; i < vertices.size(); i++)
             verticesVC.push_back({vertices[i], colors[i].Vec()});
         
         return { verticesVC, triangles };
